Refactor Entity/EntityDescriptions into base class.
