+++ homeassistant/components/fronius/sensor.py
@@ -29,7 +29,6 @@ from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import StateType
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import (
     DOMAIN,
@@ -41,11 +40,11 @@ from .const import (
     get_meter_location_description,
     get_ohmpilot_state_message,
 )
+from .entity import FroniusEntityDescription, _FroniusEntity
 
 if TYPE_CHECKING:
     from . import FroniusConfigEntry
     from .coordinator import (
-        FroniusCoordinatorBase,
         FroniusInverterUpdateCoordinator,
         FroniusLoggerUpdateCoordinator,
         FroniusMeterUpdateCoordinator,
@@ -105,7 +104,7 @@ async def async_setup_entry(
 
 
 @dataclass(frozen=True)
-class FroniusSensorEntityDescription(SensorEntityDescription):
+class FroniusSensorEntityDescription(FroniusEntityDescription, SensorEntityDescription):
     """Describes Fronius sensor entity."""
 
     default_value: StateType | None = None
@@ -695,30 +694,12 @@ STORAGE_ENTITY_DESCRIPTIONS: list[FroniusSensorEntityDescription] = [
 ]
 
 
-class _FroniusSensorEntity(CoordinatorEntity["FroniusCoordinatorBase"], SensorEntity):
+class _FroniusSensorEntity(_FroniusEntity, SensorEntity):
     """Defines a Fronius coordinator entity."""
 
     entity_description: FroniusSensorEntityDescription
 
     _attr_has_entity_name = True
-
-    def __init__(
-        self,
-        coordinator: FroniusCoordinatorBase,
-        description: FroniusSensorEntityDescription,
-        solar_net_id: str,
-    ) -> None:
-        """Set up an individual Fronius meter sensor."""
-        super().__init__(coordinator)
-        self.entity_description = description
-        self.response_key = description.response_key or description.key
-        self.solar_net_id = solar_net_id
-        self._attr_native_value = self._get_entity_value()
-        self._attr_translation_key = description.key
-
-    def _device_data(self) -> dict[str, Any]:
-        """Extract information for SolarNet device from coordinator data."""
-        return self.coordinator.data[self.solar_net_id]
 
     def _get_entity_value(self) -> Any:
         """Extract entity value from coordinator. Raises KeyError if not included in latest update."""
